+++ packages/expo-file-system/ios/EXFileSystem/EXFileSystem.h
@@ -7,7 +7,7 @@
 #import <UMFileSystemInterface/UMFileSystemInterface.h>
 #import <EXFileSystem/EXSessionResumableDownloadTaskDelegate.h>
 
-@interface EXFileSystem : UMExportedModule <UMEventEmitter, UMModuleRegistryConsumer, UMFileSystemInterface, EXResumableTaskRegister>
+@interface EXFileSystem : UMExportedModule <UMEventEmitter, UMModuleRegistryConsumer, UMFileSystemInterface, EXSessionRegister>
 
 @property (nonatomic, readonly) NSString *documentDirectory;
 @property (nonatomic, readonly) NSString *cachesDirectory;