+++ packages/expo-file-system/ios/EXFileSystem/EXSessionTasks/EXSessionDownloadTaskDelegate.h
@@ -2,22 +2,14 @@
 
 #import <EXFileSystem/EXSessionTaskDelegate.h>
 
-typedef void (^EXDownloadDelegateOnWriteCallback)(NSURLSessionDownloadTask *task, int64_t bytesWritten, int64_t totalBytesWritten, int64_t totalBytesExpectedToWrite);
-
 @interface EXSessionDownloadTaskDelegate : EXSessionTaskDelegate
             
-- (instancetype)initWithResolve:(UMPromiseResolveBlock)resolve
-                         reject:(UMPromiseRejectBlock)reject
-                   localFileUrl:(NSURL *)localFileUrl
-                      serverUrl:(NSURL *)serverUrl
-                      md5Option:(BOOL)md5Option;
-
-- (instancetype)initWithResolve:(UMPromiseResolveBlock)resolve
-                         reject:(UMPromiseRejectBlock)reject
-                   localFileUrl:(NSURL *)localFileUrl
-                      serverUrl:(NSURL *)serverUrl
-                      md5Option:(BOOL)md5Option
-                onWriteCallback:(EXDownloadDelegateOnWriteCallback)onWrite;
+- (instancetype)initWithSessionRegister:(id<EXSessionRegister>)sessionRegister
+                                resolve:(UMPromiseResolveBlock)resolve
+                                 reject:(UMPromiseRejectBlock)reject
+                           localFileUrl:(NSURL *)localFileUrl
+                              serverUrl:(NSURL *)serverUrl
+                              md5Option:(BOOL)md5Option;
 
 - (void)handleDidFinishDownloadingToURL:(NSURL *)location task:(NSURLSessionDownloadTask *)downloadTask;